Convert assembly code to NASM and remove prebuilt objects, support builds using Open Watcom tools on Linux.

+++ FASTDOOM/doomstat.h
@@ -224,7 +224,7 @@ extern ticcmd_t localcmds[BACKUPTICS];
 extern int maketic;
 extern int nettics;
 
-extern byte autorun;
+extern int autorun;
 
 extern char *mapnames[];
 extern char *mapnames2[];
